feat: complete q1 - q3 without routing libraries
