+++ client/package.json
@@ -13,9 +13,7 @@
     "axios": "^1.9.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "react-hot-toast": "^2.5.2",
-    "react-router-dom": "^7.6.2",
-    "uuid": "^11.1.0"
+    "react-hot-toast": "^2.5.2"
   },
   "devDependencies": {
     "@eslint/js": "^9.25.0",
